+++ opaque_typedef_macros/src/attrs.rs
@@ -8,7 +8,7 @@ pub fn is_attr_with_path(attr: &syn::Attribute, path: &[&str]) -> bool {
     attr.path
         .segments
         .iter()
-        .map(|seg| seg.ident.as_ref())
+        .map(|seg| &seg.ident)
         .eq(path.into_iter().map(|&s| s))
 }
 
@@ -16,9 +16,9 @@ pub fn is_attr_with_path(attr: &syn::Attribute, path: &[&str]) -> bool {
 /// Checks whether the word meta item with the given path is specified.
 pub fn has_word_meta(meta: &syn::Meta, ident_path: &[&str]) -> bool {
     match *meta {
-        syn::Meta::Word(ref ident) => ident_path.len() == 1 && ident.as_ref() == ident_path[0],
+        syn::Meta::Word(ref ident) => ident_path.len() == 1 && ident == ident_path[0],
         syn::Meta::List(ref metalist) => {
-            if ident_path.len() > 1 && metalist.ident.as_ref() == ident_path[0] {
+            if ident_path.len() > 1 && metalist.ident == ident_path[0] {
                 metalist
                     .nested
                     .iter()
@@ -43,14 +43,14 @@ pub fn get_meta_content_by_path(meta: syn::Meta, path: &[&str]) -> Vec<syn::Nest
 }
 
 
-fn append_meta_content_by_path<'a>(meta: syn::Meta, path: &[&str], vec: &mut Vec<syn::NestedMeta>) {
-    if path.len() == 0 {
+fn append_meta_content_by_path(meta: syn::Meta, path: &[&str], vec: &mut Vec<syn::NestedMeta>) {
+    if path.is_empty() {
         return;
     }
-    assert!(path.len() > 0);
+    assert!(!path.is_empty());
     match meta {
         syn::Meta::List(metalist) => {
-            if metalist.ident.as_ref() == path[0] {
+            if metalist.ident == path[0] {
                 append_meta_items_by_path(metalist.nested, &path[1..], vec);
             }
         },
@@ -59,16 +59,16 @@ fn append_meta_content_by_path<'a>(meta: syn::Meta, path: &[&str], vec: &mut Vec
 }
 
 
-fn append_meta_items_by_path<'a, I>(nested_items: I, path: &[&str], vec: &mut Vec<syn::NestedMeta>)
+fn append_meta_items_by_path<I>(nested_items: I, path: &[&str], vec: &mut Vec<syn::NestedMeta>)
 where
     I: IntoIterator<Item = syn::NestedMeta>,
 {
-    if path.len() == 0 {
+    if path.is_empty() {
         vec.extend(nested_items);
         return;
     }
-    assert!(path.len() > 0);
-    for nested_meta in nested_items.into_iter() {
+    assert!(!path.is_empty());
+    for nested_meta in nested_items {
         if let syn::NestedMeta::Meta(meta) = nested_meta {
             append_meta_content_by_path(meta, path, vec);
         }