+++ opaque_typedef_macros/src/lib.rs
@@ -4,6 +4,7 @@
 #[macro_use]
 extern crate lazy_static;
 extern crate proc_macro;
+extern crate proc_macro2;
 #[macro_use]
 extern crate quote;
 extern crate strum;
@@ -11,7 +12,7 @@ extern crate strum;
 extern crate strum_macros;
 extern crate syn;
 
-use proc_macro::TokenStream;
+use proc_macro2::TokenStream;
 use syn::DeriveInput;
 
 use type_props::{Sizedness, TypeProps};
@@ -24,24 +25,24 @@ mod utils;
 
 /// The entrypoint for a `#[derive(OpaqueTypedef)]`-ed type.
 #[proc_macro_derive(OpaqueTypedef, attributes(opaque_typedef))]
-pub fn opaque_typedef(input: TokenStream) -> TokenStream {
+pub fn opaque_typedef(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
     let input: DeriveInput = syn::parse(input).unwrap();
-    let gen = gen_opaque_typedef_impls(input, Sizedness::Sized);
+    let gen = gen_opaque_typedef_impls(&input, Sizedness::Sized);
     gen.into()
 }
 
 
 /// The entrypoint for a `#[derive(OpaqueTypedefUnsized)]`-ed type.
 #[proc_macro_derive(OpaqueTypedefUnsized, attributes(opaque_typedef))]
-pub fn opaque_typedef_unsized(input: TokenStream) -> TokenStream {
+pub fn opaque_typedef_unsized(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
     let input: DeriveInput = syn::parse(input).unwrap();
-    let gen = gen_opaque_typedef_impls(input, Sizedness::Unsized);
+    let gen = gen_opaque_typedef_impls(&input, Sizedness::Unsized);
     gen.into()
 }
 
 
 /// Generates additional impls for a `#[derive(OpaqueTypedef*)]`-ed type.
-fn gen_opaque_typedef_impls(input: DeriveInput, sizedness: Sizedness) -> quote::Tokens {
+fn gen_opaque_typedef_impls(input: &DeriveInput, sizedness: Sizedness) -> TokenStream {
     let props = TypeProps::load(&input, sizedness);
     props.gen_impls()
 }